remove unnecessary methods in LazyThen
